+++ rust/svrb/src/errors.rs
@@ -4,8 +4,6 @@
 //
 use std::fmt;
 
-use prost::DecodeError;
-
 use crate::proto::svrb;
 
 #[derive(Debug, displaydoc::Display, PartialEq)]
@@ -14,8 +12,6 @@ pub enum Error {
     BadData,
     /// Unexpected or missing server response
     BadResponse,
-    /// Response status is not OK: {0}
-    BadResponseStatus(ErrorStatus),
     /// Inputs {got} do not match the correct number of servers {servers}
     NumServers { servers: usize, got: usize },
     /// No auth version was usable.
@@ -28,23 +24,8 @@ pub enum Error {
 
 impl std::fmt::Display for svrb::response4::Status {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}", self.as_str_name())
+        write!(f, "{:?}", self)
     }
-}
-
-/// Represents an erroneous SVR3 response status
-#[derive(Debug, strum::Display, PartialEq)]
-pub enum ErrorStatus {
-    Unset,
-    Missing,
-    InvalidRequest,
-    Error,
 }
 
 impl std::error::Error for Error {}
-
-impl From<DecodeError> for Error {
-    fn from(_err: DecodeError) -> Self {
-        Self::BadData
-    }
-}